rag_cursor_sonnet7/requirements.txt
docling>=0.1.0
lancedb>=0.3.3
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0 